requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
Click==7.0
cloudpickle==1.1.1
cycler==0.10.0
dask==1.2.2
DateTime==4.3
decorator==4.4.0
enum34==1.1.6
Flask==1.0.3
ftfy==4.4.3
funcsigs==1.0.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
html5lib==1.0.1
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==1.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.2
numpy==1.16.3
pbr==5.2.0
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
stevedore==1.30.1
subprocess32==3.5.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
termcolor==1.1.0
toolz==0.9.0
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
zope.interface==4.6.0
